Wait at least at second when resending a message requiring an acknowledge
The resend loop isn't synchronised with the application, so when the
applications sends a message with an acknowledge request, it occassionaly
happens that the resend loop detects this message before one second
has passed and resends it again. This problem is now avoided by waiting
an additional second after the first message has been sent.